fix: using double instead of float in Merchant for minimum monthly fee

+++ src/main/java/dev/aleixmorgadas/merchants/Merchant.java
@@ -16,5 +16,5 @@ public class Merchant {
     String email;
     String liveOn;
     String disbursementFrequency;
-    float minimumMonthlyFee;
+    double minimumMonthlyFee;
 }
